refactor(ai): extract translation UI helpers, align batch CLI args, and update documentation
This commit refactors the AI translation pipeline to unify behavior between the batch and single-passage translation tools, eliminates duplicated code, enables custom provider selection for batch processing, and updates the directory documentation to reflect these changes.

Key changes:
- Created `exporter/analysis/ui_utils.py`: Extracted and consolidated progress reporting (`_print_translation_progress`), raw response log formatting (`_build_raw_responses_log`), and debug artifact writing (`_write_ai_debug_artifacts`). This ensures that CLI feedback and diagnostics are identical across both batch and interactive runs.
- Updated `exporter/analysis/study_passage.py`: Removed local implementations of the extracted helpers and imported them from `ui_utils.py`, maintaining existing functionality.
- Enhanced `exporter/analysis/ai_batch_translate.py`:
  - Added CLI arguments (`--provider`, `--model`, `--debug`) and validation to ensure model and provider are supplied together.
  - Passed parameters down to `translate_sentence()` inside the verse processing loop.
  - Enabled writing verse-specific debug files to output directories upon verse failure or when verbose logging is enabled.
- Updated `tests/exporter/analysis/test_study_passage.py`: Updated test imports to reference the new `ui_utils` functions and verified the new argument parsing constraints.
- Updated `exporter/analysis/README.md`:
  - Documented the new `ui_utils.py` helper module in the "Support" file map.
  - Added CLI usage examples for the new `--provider`, `--model`, and `--debug` parameters to the "Batch Verse Workflow" section.
